Move relationships components in pages/people/relationships/*
