--- components/Relationships/RelationshipSearchTable.module.scss
@@ -1,7 +1,0 @@
-.restrictedRow {
-  background-color: #f8f8f8;
-}
-
-.uppercase {
-  text-transform: uppercase;
-}
